Cleanup Models + Migrations + Seeders (#1)
I have updated the models to match the schema diagram that I created. I have also addressed some issues with the existing files that I encountered. As a result I regenerated the initial migration and the seeders. I tested the migration and seed process against a local sqlite3 database as well as a postgresql database to make sure both are happy. We might want to review the to_dict function in the models to see if we always want to return all columns. Also, I included the group and group members tables however their usage is optional/ not required.